fix duplicate react sharing
