+++ nextjs-react/host-app/async-pages/_app.js
@@ -1,0 +1,5 @@
+function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}
+
+export default MyApp;
