+++ nextjs-react/host-app/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@module-federation/nextjs-mf": "^4.0.6",
+    "@module-federation/nextjs-mf": "^4.1.0",
     "next": "12.2.2",
     "react": "17.0.2",
     "react-dom": "17.0.2"
