+++ nextjs-react/host-app/pages/_app.js
@@ -1,8 +1,15 @@
 import '@module-federation/nextjs-mf/beta/include-defaults'
 import '../styles/globals.css';
+import dynamic from 'next/dynamic';
+console.log(__webpack_share_scopes__)
+const App = dynamic(async () => {
+  return import('../async-pages/_app')
+}, {
+  ssr: false,
+});
 
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+function MyApp(props) {
+  return <App {...props} />;
 }
 
 export default MyApp;
