+++ nextjs-react/host-app/pages/index.js
@@ -1,16 +1,11 @@
-import Button from '../components/Button';
 import dynamic from 'next/dynamic';
-const RemoteButton = dynamic(() => import('remote/Button'), {
+const Page = dynamic(async () => {
+  return import('../async-pages/index')
+}, {
   ssr: false,
 });
 export default function Home() {
   return (
-    <div style={{ padding: '2%' }}>
-      <h1>Next JS and React</h1>
-      <h2>Host - Button</h2>
-      <Button />
-      <h2>Client - Button</h2>
-      <RemoteButton />
-    </div>
+    <Page/>
   );
 }
